requirements.txt
faiss-cpu
numpy
sentence-transformers
rich
pydantic
sounddevice
soundfile
edge-tts
faster-whisper